+++ r5dev/vscript/languages/squirrel_re/vsquirrel.h
@@ -43,6 +43,10 @@ private:
 };
 #pragma pack(pop)
 
+extern void(*ServerScriptRegister_Callback)(CSquirrelVM* s);
+extern void(*ClientScriptRegister_Callback)(CSquirrelVM* s);
+extern void(*UiScriptRegister_Callback)(CSquirrelVM* s);
+
 inline CMemory p_CSquirrelVM_Init;
 inline bool(*v_CSquirrelVM_Init)(CSquirrelVM* s, SQCONTEXT context, SQFloat curtime);
